move utils.kernel (formerly entry_point and lib.kernel) to kernel.util
