+++ IPython/kernel/tests/test_util.py
