+++ IPython/kernel/util.py
