+++ IPython/lib/kernel.py
@@ -1,12 +1,12 @@
 """[DEPRECATED] Utilities for connecting to kernels
 
-Moved to IPython.utils.kernel, where it always belonged.
+Moved to IPython.kernel.util
 """
 
 import warnings
-warnings.warn("IPython.lib.kernel moved to IPython.utils.kernel in IPython 0.14",
+warnings.warn("IPython.lib.kernel moved to IPython.kernel in IPython 0.14",
     DeprecationWarning
 )
 
-from IPython.utils.kernel import *
+from IPython.kernel.util import *
