+++ IPython/zmq/kernelapp.py
@@ -40,7 +40,7 @@ from IPython.utils.traitlets import (
     DottedObjectName,
 )
 from IPython.utils.importstring import import_item
-from IPython.utils.kernel import write_connection_file
+from IPython.kernel import write_connection_file
 # local imports
 from IPython.zmq.heartbeat import Heartbeat
 from IPython.zmq.parentpoller import ParentPollerUnix, ParentPollerWindows
